+++ index.js
@@ -1,6 +1,1 @@
-const faparser = require('./src/faparser')
-
-module.exports = { 
-    getFilm: faparser.film, 
-    search: faparser.search
-}
+export { search, film, types } from './src/index.js';
